display and delete blueprints

+++ src/Aco228.GoogleServices/Extensions/BucketFileExtensions.cs
@@ -6,27 +6,31 @@ namespace Aco228.GoogleServices.Extensions;
 
 public static class BucketFileExtensions
 {
-    public static string GetUrl(this BucketFile bucketFile)
+    public static string GetUrl(this BucketFile? bucketFile)
     {
+        if (bucketFile == null) return string.Empty;
         return $"https://storage.googleapis.com/{bucketFile.BucketName}/{bucketFile.FileName}";
     }
     
-    public static Task<string> GetTextAsync(this BucketFile bucketFile)
+    public static Task<string> GetTextAsync(this BucketFile? bucketFile)
     {
+        if (bucketFile == null) return Task.FromResult(string.Empty);
         var httpClientFactory = ServiceProviderHelper.GetService<IHttpClientFactory>()!;
         var client = httpClientFactory.CreateClient();
         return client.GetStringAsync($"https://storage.googleapis.com/{bucketFile.BucketName}/{bucketFile.FileName}");
     }
     
-    public static Task<byte[]> GetBinaryAsync(this BucketFile bucketFile)
+    public static Task<byte[]> GetBinaryAsync(this BucketFile? bucketFile)
     {
+        if (bucketFile == null) return Task.FromResult(Array.Empty<byte>());
         var httpClientFactory = ServiceProviderHelper.GetService<IHttpClientFactory>()!;
         var client = httpClientFactory.CreateClient();
         return client.GetByteArrayAsync($"https://storage.googleapis.com/{bucketFile.BucketName}/{bucketFile.FileName}");
     }
 
-    public static async Task<bool> DeleteAsync(this BucketFile bucketFile)
+    public static async Task<bool> DeleteAsync(this BucketFile? bucketFile)
     {
+        if (bucketFile == null) return false;
         var bucket = DefaultGoogleBucket.Construct(bucketFile.BucketName);
         return await bucket.DeleteFile(bucketFile);
     }
